Deleted integrated pattern document
Since integrating fixes from USGS in e583a9f, the integrated pattern
document is no longer a representative state of the project. All
refinement of symbols should focus on the individual svg files in the
assets/svg directory.

TODO: we should create some sort of "rollup" to
package the symbols as a single SVG for certain uses.